% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_png.R
\name{lossy_choices}
\alias{lossy_choices}
\title{Get available lossy optimization methods}
\usage{
lossy_choices()
}
\value{
A list containing vectors \code{optimizer} and \code{ditherer}, plus \code{default}.
}
\description{
Return available optimizer and ditherer choices for \code{tinypng(lossy = ...)}.
}
