% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyimg-package.R
\docType{package}
\name{tinyimg-package}
\alias{tinyimg}
\alias{tinyimg-package}
\title{tinyimg: Optimize and Compress Images}
\description{
The tinyimg package provides tools for optimizing and compressing images
using Rust libraries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yihui/tinyimg}
  \item Report bugs at \url{https://github.com/yihui/tinyimg/issues}
}

}
\author{
\strong{Maintainer}: Yihui Xie \email{xie@yihui.name} (\href{https://orcid.org/0000-0003-0645-5666}{ORCID}) (https://yihui.org)

Other contributors:
\itemize{
  \item Authors of the dependency Rust crates (see AUTHORS file) [contributor]
}

}
\keyword{internal}
